Improved projectiles and timeout is working

+++ src/tools/vectorTools.go
@@ -9,8 +9,8 @@ func GetMagnitude(vec *rl.Vector2) float32 {
   return float32(math.Sqrt(math.Pow(float64(vec.X), 2) + math.Pow(float64(vec.Y), 2)))
 }
 
-func GetXYComponent(magnitude, angle float64) rl.Vector2 {
-  return rl.NewVector2(float32(magnitude*math.Sin(angle)), float32(magnitude*math.Cos(angle)))
+func GetXYComponent(magnitude, angle float32) rl.Vector2 {
+  return rl.NewVector2(magnitude*float32(math.Sin(float64(angle))), magnitude*float32(math.Cos(float64(angle))))
 }
 
 func GetAngle(vec rl.Vector2) float32 {
